rp/requirements.txt
pyserial==3.5
matplotlib==3.8.0
numpy==1.24.3
